+++ src/bonuses/spells/spell-focus.mjs
@@ -198,7 +198,7 @@ Hooks.on(customGlobalHooks.itemGetTypeChatData, (
 
 // register hint on focused spell
 registerItemHint((hintcls, actor, item, _data) => {
-    if (!(item instanceof pf1.documents.item.ItemSpellPF)) {
+    if (!(item instanceof pf1.documents.item.ItemSpellPF) || !actor) {
         return;
     }